feat(keychain): refactoring and exploding
adjustements in the keychain implementation in progress
